requirements.txt
torch
transformers
streamlit
pdfminer.six
tqdm
